+++ tests/testthat/test-utility.R
@@ -5,6 +5,14 @@ responses <- list(Q1 = 3:5, Q2a = 1:3, Q2b = 1:3, Q3a = 1:3, Q3b = 1:3, Q4a = 1:
                   Q4b = 1:3, Q5 = 1:3, Q6a = 1:3, Q6b = 1:3, Q6c = 1:3, Q7 = 1:3)
 sf12 <- rlang::exec(SF12, !!!responses)
 
+responses_sf36 <- list(
+  Q1 = 1:2, Q2 = 1:2, Q3 = 1:2, Q4 = 1:2, Q5 = 1:2, Q6 = 1:2, Q7 = 1:2, Q8 = 1:2, Q9 = 1:2,
+  Q10 = 1:2, Q11 = 1:2, Q12 = 1:2, Q13 = 1:2, Q14 = 1:2, Q15 = 1:2, Q16 = 1:2, Q17 = 1:2, Q18 = 1:2,
+  Q19 = 1:2, Q20 = 1:2, Q21 = 1:2, Q22 = 1:2, Q23 = 1:2, Q24 = 1:2, Q25 = 1:2, Q26 = 1:2, Q27 = 1:2,
+  Q28 = 1:2, Q29 = 1:2, Q30 = 1:2, Q31 = 1:2, Q32 = 1:2, Q33 = 1:2, Q34 = 1:2, Q35 = 1:2, Q36 = 1:2
+)
+sf36 <- rlang::exec(SF36, !!!responses_sf36)
+
 test_that("uk works as expected", {
   expect_equal(uk(sf6d), c(0.937, 0.660, 0.570))
 })
@@ -15,6 +23,7 @@ test_that("utility works as expected", {
 
 test_that("utility gives appropriate error messages", {
   expect_error(utility(sf12), class = "SF6Dvalues_error_not_SF6D")
+  expect_error(utility(sf36), class = "SF6Dvalues_error_not_SF6D")
   expect_error(utility(levels), class = "SF6Dvalues_error_not_SF6D")
   expect_error(utility(1), class = "SF6Dvalues_error_not_SF6D")
   expect_error(utility("111111"), class = "SF6Dvalues_error_not_SF6D")